fix nil item look up when scoped by a list
Command.search_list_for_item was incorrectly using first instead of
find. This caused it to always return the first item in the list
if you searched for a name within that list that didn't exist, or
to throw an exception if the list was empty (due to the returned
item being nil.) It now looks for the correct item and outputs an
error message if no matching item is found in the list.